telegram_bot/requirements.txt
python-telegram-bot==20.6
requests
web3
python-dotenv
